refactor(settings): extract connection-detail controller + enabled-model manager (arch R7)
provider-connection-detail.tsx was the renderer's densest state cluster
(983 lines, 12 useState in ConnectionDetailInner + a shared keyed action
guard + lifecycle gate). Split following the settings house pattern:

- use-connection-detail.ts (controller hook, ~520 lines): owns the whole
  mutually-exclusive action state machine (save / test / fetch-models /
  save-enabled-models / set-default / delete, all gated through one
  useKeyedActionGuard), the credential-presence probe, the prop-sync
  effects, every derived flag, and the pure snapshot/equality helpers +
  oauthLoginServiceFor. Returns a controller object; ZERO behavior change,
  every identifier and statement moved verbatim.
- provider-enabled-model-manager.tsx (~187 lines): the roving-tabindex
  model-list editor (owns query + activeRowId) — an independent cluster.
- provider-connection-detail.tsx (983 -> 373 lines): thin view that
  destructures the hook, plus ConnectionEndpointField / the OAuth &
  GitHub Copilot re-login notices (kept here as presentation).

The connection/model/test/delete sub-clusters stay in one controller on
purpose: they share a single keyed-action-guard state machine (mutual
exclusion) and one lifecycle/current-ness gate, and cross-call (save
auto-fetches models) — splitting them further would thread the guard and
risk behavior drift.
